몬스터 메모리풀 적용
언제 무슨 버그로 터질지 모름 테스트 엄청 많이 해보기

+++ VOODOODOLL/VooDoo_Doll/protocol.h
@@ -76,7 +76,7 @@ struct SC_MOVE_PLAYER_PACKET {
 	char	type;
 	short	id;
 	short	character_num;
-	short	HP;
+	float	HP;
 	XMFLOAT3 Look, Up, Right, Pos;
 	// XMFLOAT3 BulletPos;
 	bool overwrite;
